magento-engcom/community-portal#235 Build Project Recommendation Engine
- refactoring and integration
