+++ functions/src/controllers/ProjectRecommendationController/ProjectRecommendationController.ts
@@ -1,12 +1,18 @@
+import Counter from './../../custom/Counter';
+
 interface ProjectTrafficControllerInterface {
   getRecommendations(data: any): (result: any) => any;
 }
 
-function getSecondLevelCount(map: Map<string, number>, firstKey: string, secondKey: string) {
+function getSecondLevelCount(map: Map<string, any>, firstKey: string, secondKey: string) {
     if (map.has(firstKey)) {
         if (map.get(firstKey).has(secondKey)) return map.get(firstKey).get(secondKey);
     }
     return 0;
+}
+
+function getSignature(array: string[]) {
+  return JSON.stringify(array);
 }
 
 function getStateFromSignature(signature: string) {
@@ -19,12 +25,40 @@ function getNextStateSignature(signature: string, next: string) {
   return getSignature(nextState);
 }
 
-function delta(project_seq_signature: string, next_proj: string, hitTransitionFunc: any, missTransitionFunc: any, values: Counter) {
-  return (hitTransitionFunc(project_seq_signature, next_proj) - missTransitionFunc(project_seq_signature, next_proj)) * values.get(getNextStateSignature(project_seq_signature, next_proj));
+function delta(
+  signature: string,
+  nextProj: string,
+  hitTransitionMap: any,
+  missTransitionMap: any,
+  values: Counter,
+) {
+  return (
+    getSecondLevelCount(hitTransitionMap, signature, nextProj)
+      - getSecondLevelCount(missTransitionMap, signature, nextProj)
+  )
+    * values.get(getNextStateSignature(signature, nextProj));
 }
 
-hitTransitionFunc: (signature: string, next: string) => getSecondLevelCount(hitTransitionMap, signature, next),
-missTransitionFunc: (signature: string, next: string) => getSecondLevelCount(missTransitionMap, signature, next),
+// k represents the number of projects to recommend
+function packageModel(
+  values: Counter,
+  hitTransitionMap: any,
+  missTransitionMap: any,
+  observedTransitions: any,
+  k: number,
+) {
+  return (state: string[], nextProj: string) => {
+    const signature = getSignature(state);
+    if (!observedTransitions.has(signature)) return [];
+    const nextProjects = observedTransitions.get(signature);
+    const nextStates = nextProjects.map((nextProj: string) => ({
+      delta: delta(signature, nextProj, hitTransitionMap, missTransitionMap, values),
+      id: nextProj,
+    }));
+    nextStates.sort((a: any, b: any) => a.delta < b.delta);
+    return nextStates.slice(0, k).map((project: any) => project.id);
+  }
+}
 
 export default class ProjectRecommendationController implements ProjectTrafficControllerInterface {
 
@@ -32,10 +66,12 @@ export default class ProjectRecommendationController implements ProjectTrafficCo
   getModel(data: any) {
     return (result: any) => {
       const { values, hitTransitionMap, missTransitionMap, observedTransitions } = JSON.parse(result.Body.toString());
+      const model = packageModel(values, hitTransitionMap, missTransitionMap, observedTransitions, 3);
 
       return { model };
     };
   }
+
   getRecommendations(data: any) {
     return (result: any) => {
       const { recommended } = result;
